Fix NPM package upload
